src/test/java/com/bbringworld/ezparkapi/domain/spot/application/service/SpotServiceTest.java
package com.bbringworld.ezparkapi.domain.spot.application.service;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;

import static org.assertj.core.api.FactoryBasedNavigableListAssert.assertThat;
import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
class SpotServiceTest {

    @Autowired
    SpotService spotService;

    @Test
    @DisplayName("search by id")
    void search_by_id() {
        assertNotNull(spotService.getById(1L));
    }
}